requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.7.2
    # via django
certifi==2023.5.7
    # via requests
charset-normalizer==3.1.0
    # via requests
cmake==3.26.4
    # via triton
django==4.2.2
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.14.0
    # via -r requirements.in
faiss-cpu==1.7.4
    # via txtai
filelock==3.12.2
    # via
    #   huggingface-hub
    #   torch
    #   transformers
    #   triton
fsspec==2023.6.0
    # via huggingface-hub
gunicorn==20.1.0
    # via -r requirements.in
huggingface-hub==0.15.1
    # via transformers
idna==3.4
    # via requests
jinja2==3.1.2
    # via torch
lit==16.0.6
    # via triton
markupsafe==2.1.3
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.1
    # via torch
numpy==1.25.0
    # via
    #   transformers
    #   txtai
nvidia-cublas-cu11==11.10.3.66
    # via
    #   nvidia-cudnn-cu11
    #   nvidia-cusolver-cu11
    #   torch
nvidia-cuda-cupti-cu11==11.7.101
    # via torch
nvidia-cuda-nvrtc-cu11==11.7.99
    # via torch
nvidia-cuda-runtime-cu11==11.7.99
    # via torch
nvidia-cudnn-cu11==8.5.0.96
    # via torch
nvidia-cufft-cu11==10.9.0.58
    # via torch
nvidia-curand-cu11==10.2.10.91
    # via torch
nvidia-cusolver-cu11==11.4.0.1
    # via torch
nvidia-cusparse-cu11==11.7.4.91
    # via torch
nvidia-nccl-cu11==2.14.3
    # via torch
nvidia-nvtx-cu11==11.7.91
    # via torch
packaging==23.1
    # via
    #   huggingface-hub
    #   transformers
pytz==2023.3
    # via djangorestframework
pyyaml==6.0
    # via
    #   huggingface-hub
    #   transformers
    #   txtai
regex==2023.6.3
    # via transformers
requests==2.31.0
    # via
    #   huggingface-hub
    #   transformers
safetensors==0.3.1
    # via transformers
sqlparse==0.4.4
    # via django
sympy==1.12
    # via torch
tokenizers==0.13.3
    # via transformers
torch==2.0.1
    # via
    #   triton
    #   txtai
tqdm==4.65.0
    # via
    #   huggingface-hub
    #   transformers
transformers==4.30.2
    # via txtai
triton==2.0.0
    # via torch
txtai==5.5.1
    # via -r requirements.in
typing-extensions==4.7.0
    # via
    #   asgiref
    #   huggingface-hub
    #   torch
urllib3==2.0.3
    # via requests
wheel==0.40.0
    # via
    #   nvidia-cublas-cu11
    #   nvidia-cuda-cupti-cu11
    #   nvidia-cuda-runtime-cu11
    #   nvidia-curand-cu11
    #   nvidia-cusparse-cu11
    #   nvidia-nvtx-cu11

# The following packages are considered to be unsafe in a requirements file:
# setuptools
